move cmd runner to util.runtime
