+++ components/server/src/main/java/ome/smuggler/core/service/omero/impl/OmeCliCommandRunner.java
@@ -7,10 +7,10 @@ import java.nio.file.Path;
 import java.util.Optional;
 import java.util.function.Consumer;
 
-import ome.smuggler.core.io.CommandRunner;
 import util.io.StreamOps;
 import util.lambda.FunctionE;
 import util.object.Pair;
+import util.runtime.CommandRunner;
 
 /**
  * Spawns a separate JVM process to run an OME CLI command.
