+++ components/util/src/test/java/util/runtime/CommandRunnerTest.java
@@ -1,14 +1,10 @@
-package integration.cli;
+package util.runtime;
 
 import static org.junit.Assert.*;
 
-import ome.smuggler.core.io.CommandRunner;
 import util.io.StreamOps;
 import org.junit.Test;
 import util.object.Pair;
-import util.runtime.BaseProgramArgument;
-import util.runtime.CommandBuilder;
-import util.runtime.ProgramArgument;
 import util.runtime.jvm.*;
 
 import java.io.InputStream;
