+++ components/util/src/test/java/util/runtime/LinesProducer.java
@@ -1,4 +1,4 @@
-package integration.cli;
+package util.runtime;
 
 import java.util.function.Consumer;
